+++ src/game_object.cpp
@@ -1,30 +1,30 @@
 #include "game_object.hpp"
 
-GameObject::GameObject(double pX, double pY): 
+GameObject::GameObject(float pX, float pY): 
     pos(pX,pY) {}
 
 
-void GameObject::setPosX(double x) {this->pos.x = x;}
-void GameObject::setPosY(double y) {this->pos.y = y;}
+void GameObject::setPosX(float x) {this->pos.x = x;}
+void GameObject::setPosY(float y) {this->pos.y = y;}
 
 Point GameObject::getPos() {return this->pos;}
-double GameObject::getPosX() {return this->pos.x;}
-double GameObject::getPosY() {return this->pos.y;}
+float GameObject::getPosX() {return this->pos.x;}
+float GameObject::getPosY() {return this->pos.y;}
 
-Drawable::Drawable(double x, double y, double sX, double sY):
+Drawable::Drawable(float x, float y, float sX, float sY):
     GameObject::GameObject(x,y), speedX(sX), speedY(sY) {}
-Drawable::Drawable(double x, double y):
+Drawable::Drawable(float x, float y):
     Drawable::Drawable(x,y,0,0) {}
 
-double Drawable::getSpeedX() {return this->speedX;}
-double Drawable::getSpeedY() {return this->speedY;}
+float Drawable::getSpeedX() {return this->speedX;}
+float Drawable::getSpeedY() {return this->speedY;}
 
-void Drawable::setSpeedX(double x) {this->speedX = x;}
-void Drawable::setSpeedY(double y) {this->speedY = y;}
+void Drawable::setSpeedX(float x) {this->speedX = x;}
+void Drawable::setSpeedY(float y) {this->speedY = y;}
 
-void Drawable::addSpeedX(double x) {
+void Drawable::addSpeedX(float x) {
     this->speedX += x;
 }
-void Drawable::addSpeedY(double y) {
+void Drawable::addSpeedY(float y) {
     this->speedY += y;
 }
