+++ src/hitbox.cpp
@@ -1,6 +1,6 @@
 #include "hitbox.hpp"
 
-Hitbox::Hitbox(double pX, double pY, hitboxType t):
+Hitbox::Hitbox(float pX, float pY, hitboxType t):
     GameObject::GameObject(pX,pY), type(t) {}
 
 Hitbox::~Hitbox() {}
@@ -20,8 +20,8 @@ bool Hitbox::updatePosition() {
 
 void Hitbox::setTarget(Drawable *d) {this->target = d;}
 
-RectangleHitbox::RectangleHitbox(double pX, double pY,
-                                double w, double h, hitboxType t):
+RectangleHitbox::RectangleHitbox(float pX, float pY,
+                                float w, float h, hitboxType t):
     Hitbox::Hitbox(pX,pY,t), width(w), height(h) {
     
     vertex[0] = {this->getPosX() - width/2, this->getPosY() - height/2};
@@ -32,11 +32,11 @@ bool RectangleHitbox::checkColision() {
     return false;
 } //placeholder
 
-double RectangleHitbox::getWidth() {return this->width;}
-double RectangleHitbox::getHeight() {return this->height;}
+float RectangleHitbox::getWidth() {return this->width;}
+float RectangleHitbox::getHeight() {return this->height;}
 Point RectangleHitbox::getVertex0() {return this->vertex[0];}
 Point RectangleHitbox::getVertex1() {return this->vertex[1];}
 
-CircleHitbox::CircleHitbox(double pX, double pY, 
-                                double r, hitboxType t):
+CircleHitbox::CircleHitbox(float pX, float pY, 
+                                float r, hitboxType t):
     Hitbox::Hitbox(pX,pY,t), radius(r) {}
